refactor(logistics): move DB persistence to infrastructure
