+++ server/features/logistics/infrastructure/README.md
@@ -1,7 +1,8 @@
 # Logistics · infrastructure (implementación de puertos)
 
-> Capa **infrastructure** del contexto Logistics. **Contiene código** desde M02b:
-> un adaptador transitorio de DB para el breach de SLA.
+> Capa **infrastructure** del contexto Logistics. **Contiene código** desde M02b
+> (adaptador de DB para el breach de SLA) y, desde **M12 — implementado /
+> pendiente de merge**, la **persistencia canónica** del contexto.
 > Ver la frontera del contexto en [`../README.md`](../README.md) y el contrato en
 > [ARCH-2](../../../../docs/architecture/backend-boundary-adr.md).
 
@@ -20,31 +21,45 @@ de persistencia y el I/O real del contexto.
 
 ## Qué vive aquí
 
-- **`sla-breach-db.ts`** (M02b) — adaptador **transitorio** que cablea el núcleo puro
-  `markOverdueSlaBreaches` (importado del barrel de `domain`) con la persistencia real
-  de `server/db-logistics.ts`. Conserva el import dinámico (lazy) de
-  `markOverdueActiveClinicSlaInstancesBreached` y expone `markOverdueSlaBreachesWithDb`
-  sin duplicar la lógica de negocio: es el único lugar del contexto que conoce el
-  `db-*`. Su existencia **no** implica mover `server/db-logistics.ts` ni adelantar
-  M12; cuando M12 formalice el repositorio sobre puertos, este adaptador se reemplaza
-  por esa infraestructura.
+- **`db-logistics.ts`** (M12) — **implementación canónica** de la persistencia del
+  contexto: **1.291 LOC** medidos en HEAD `101731d` antes del move (la cifra real;
+  el inventario ARCH-3 citaba 1.322/1.295 LOC, desactualizados). El archivo se movió
+  **completo** desde `server/db-logistics.ts`, sin reorganizar funciones, sin
+  dividirlo, sin renombrar exports y **sin reparticionar transacciones**: conserva
+  exactamente los **7 call-sites `db.transaction(`** del baseline R0, verificado por
+  `test/architecture/logistics-infrastructure-boundary-guard.test.ts` y por
+  `test/unit/infrastructure/logistics/logistics-db.test.ts`. Lo único que cambia son
+  los tres specifiers exigidos por la nueva profundidad: `../../../db.ts`,
+  `../../../../drizzle/schema.ts` y `../domain/index.ts` (mismo `server/db.ts`,
+  mismo `drizzle/schema.ts`, mismo barrel de dominio).
+- **`sla-breach-db.ts`** (M02b, reapuntado en M12) — adaptador que cablea el núcleo
+  puro `markOverdueSlaBreaches` (importado del barrel de `domain`) con la
+  persistencia real. Desde M12 consume el **archivo canónico de su propia capa**
+  (`./db-logistics.ts`) en vez del shim raíz, conservando el import dinámico (lazy)
+  de `markOverdueActiveClinicSlaInstancesBreached`, la delegación al núcleo de
+  dominio y las signatures.
+
+## Shim de compatibilidad en el root
+
+`server/db-logistics.ts` queda como **shim temporal**: sólo re-exporta la superficie
+pública desde `./features/logistics/infrastructure/db-logistics.ts`. No importa
+Drizzle, ni `drizzle/schema.ts`, ni `server/db.ts`; no contiene funciones, queries ni
+transacciones, y no declara default export. Existe porque las rutas legacy
+`server/routes/logistics-*.fastify.ts` siguen importando `../db-logistics.ts`
+(estática y dinámicamente) hasta **M14–M16**; al cerrarse esa secuencia, el shim
+desaparece.
+
+**Sin cambios de schema ni de migraciones**: M12 no toca `drizzle/schema.ts`,
+`drizzle/**`, `migrations/**`, endpoints ni contratos HTTP.
 
 ## Qué vivirá aquí (futuro, no ahora)
 
-La Fase A (cerrada en M05) **no** amplía esta capa: el único código que vive aquí es
-el adaptador transitorio de SLA. La infraestructura completa del contexto llega
-después, en su propia secuencia:
-
-- **M12 — move completo de `server/db-logistics.ts`** (~1.322 LOC) → repositorio en
-  `infrastructure/`, con las transacciones intactas y un shim documentado. Es el
-  único `db-*` que ya delega en helpers de dominio (`time-window`, `route-planning`).
 - **M13 — cache adapter** para `server/lib/logistics-route-plans-cache.ts` (TTL/keys
-  preservados; contract-test de runtime existente).
-
-Ninguno de esos milestones se adelanta aquí. **`db-logistics.ts` y la cache siguen
-siendo legacy y fuera de alcance hasta M12/M13.**
+  preservados; contract-test de runtime existente). **Pendiente**: M12 no lo adelanta.
 
 ## Qué NO hacer
 
-No mover `server/db-logistics.ts` ni la cache todavía. No crear adaptadores vacíos.
-No tocar `drizzle/schema.ts` ni migraciones.
+No mover la cache todavía (M13). No crear adaptadores vacíos. No tocar
+`drizzle/schema.ts` ni migraciones. No reparticionar las transacciones del archivo
+canónico. No importar `application`, `routes` ni Fastify desde esta capa (el guard
+de frontera lo bloquea).
